+++ packages/code-du-travail-frontend/pages/fiche-service-public/[slug].js
@@ -1,7 +1,6 @@
 import { Container, Section } from "@socialgouv/cdtn-ui";
 import { FicheServicePublic } from "@socialgouv/react-fiche-service-public";
 import getConfig from "next/config";
-import Head from "next/head";
 import { withRouter } from "next/router";
 import React from "react";
 
@@ -47,10 +46,11 @@ class Fiche extends React.Component {
     } = data;
     return (
       <Layout>
-        <Metas title={title} description={description} />
-        <Head>
-          <link rel="canonical" href={url} />
-        </Head>
+        <Metas
+          title={title}
+          description={description}
+          overrideCanonical={url}
+        />
         <Answer
           title={title}
           relatedItems={relatedItems}