+++ packages/code-du-travail-modeles/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.53.1](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.53.0...v4.53.1) (2021-09-24)
+
+**Note:** Version bump only for package @socialgouv/modeles-social
+
+
+
+
+
+# 4.53.0 (2021-09-17)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @socialgouv/fiches-travail-data to ^4.181.0 ([#3638](https://github.com/SocialGouv/code-du-travail-numerique/issues/3638)) ([6a97857](https://github.com/SocialGouv/code-du-travail-numerique/commit/6a97857d35b493b11ab98fffe990c6087c471de3))
+* **deps:** update dependency @types/jest to v27 ([#3711](https://github.com/SocialGouv/code-du-travail-numerique/issues/3711)) ([503b9e8](https://github.com/SocialGouv/code-du-travail-numerique/commit/503b9e85c4284d09fb4ad4819a4993b9cd8a964e))
+
+
+
+
+
 ## [4.52.1](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.52.0...v4.52.1) (2021-07-20)